# Asset-Manager/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Farm By Gaavwala ecommerce API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /wishlist:
    get:
      operationId: getWishlist
      tags: [wishlist]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WishlistItem"
        "401":
          description: Unauthorized
    post:
      operationId: addToWishlist
      tags: [wishlist]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WishlistInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WishlistItem"
        "401":
          description: Unauthorized
  /wishlist/{productId}:
    delete:
      operationId: removeFromWishlist
      tags: [wishlist]
      security:
        - bearerAuth: []
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed
        "401":
          description: Unauthorized
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Product:
      type: object
      required: [id, name, slug, price, stock, rating, reviewCount, featured, active]
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: integer
        originalPrice:
          type: ["integer", "null"]
        discount:
          type: ["integer", "null"]
        stock:
          type: integer
        categoryId:
          type: ["integer", "null"]
        rating:
          type: number
        reviewCount:
          type: integer
        images:
          type: array
          items:
            type: string
        badge:
          type: ["string", "null"]
        emoji:
          type: ["string", "null"]
        featured:
          type: boolean
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
...
    WishlistItem:
      type: object
      required: [id, userId, productId, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        productId:
          type: integer
        product:
          oneOf:
            - $ref: "#/components/schemas/Product"
            - type: "null"
        createdAt:
          type: string
          format: date-time
    WishlistInput:
      type: object
      required: [productId]
      properties:
        productId:
          type: integer
...